JSONUtils fix, CustomAdapter and CustomView for Recipies added, recyclerview populated

+++ app/src/main/java/com/example/gismo/chefsteps/network/model/Ingredient.java
@@ -11,7 +11,7 @@ public class Ingredient {
 
     @SerializedName("quantity")
     @Expose
-    private int quantity;
+    private double quantity;
     @SerializedName("measure")
     @Expose
     private String measure;
@@ -19,7 +19,7 @@ public class Ingredient {
     @Expose
     private String ingredient;
 
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
